Removed Angelscript namespace
Incorrect namespace scoping in the addon scriptstdstring.cpp has left some errors. To avoid dirtying the subproject, might as well remove the namespace altogether. It might have not been neccesary in the first place because the naming convention of the library already includes "as" at the beginning of most of its stuff.