Refactored Asset Management
The original approach didn't work at all as it lead to resource management being more complicated than it should be. In the original, resources _owned_ assets which was bothersome to work with so, in this refactor, I flipped it around resulting in something a lot more easier to understand and maintain.