Fixed pointers being invalidated when adding/removing components
This made it difficult for a system to add/remove components while it's currently looping though a set of components because some of its references may be invalidated at any point without any form of predictability. I opted for chunked storage to deal with adding components (using a deque, temporarily) and for removal, components are just marked "unused" and can be reused later. Components are rarely, if ever, deleted.