Removed system's dependency on layer
Layer is now stored in scene as a deque and pulled it out of unique_ptr.
physics_world will need some work. b2World will need to be pulled out of the system because it is not copyable (unfortunately) but I might have ended up doing this anyway interactactivity between layers to some degree, such as tilemap.